+++ src/gui/EntityListGui.h
@@ -7,7 +7,7 @@
 class EntityListGui : public GuiWindow {
 public:
 	EntityListGui(Gui &parent, TurnManager &managerTurns, sf::Font &textFont, unsigned int textSize, sf::Vector2f position, sf::Vector2f size, sf::Vector2f origin,
-		sf::Color backgroundColour, sf::Color hoverBackgroundColour, sf::Color borderColour, int borderThickness);
+		sf::Color backgroundColour, sf::Color hoverBackgroundColour, sf::Color borderColour, float borderThickness);
 
 	virtual void update(Input &input) override;
